% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_associate_software_token}
\alias{cognitoidentityprovider_associate_software_token}
\title{Returns a unique generated shared secret key code for the user account}
\usage{
cognitoidentityprovider_associate_software_token(AccessToken, Session)
}
\arguments{
\item{AccessToken}{The access token.}

\item{Session}{The session which should be passed both ways in challenge-response calls
to the service. This allows authentication of the user as part of the
MFA setup process.}
}
\description{
Returns a unique generated shared secret key code for the user account.
The request takes an access token or a session string, but not both.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$associate_software_token(
  AccessToken = "string",
  Session = "string"
)
}
}

\keyword{internal}
