% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_risk_configuration}
\alias{cognitoidentityprovider_describe_risk_configuration}
\title{Describes the risk configuration}
\usage{
cognitoidentityprovider_describe_risk_configuration(UserPoolId,
  ClientId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{ClientId}{The app client ID.}
}
\description{
Describes the risk configuration.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$describe_risk_configuration(
  UserPoolId = "string",
  ClientId = "string"
)
}
}

\keyword{internal}
