% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_device}
\alias{cognitoidentityprovider_get_device}
\title{Gets the device}
\usage{
cognitoidentityprovider_get_device(DeviceKey, AccessToken)
}
\arguments{
\item{DeviceKey}{[required] The device key.}

\item{AccessToken}{The access token.}
}
\description{
Gets the device.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$get_device(
  DeviceKey = "string",
  AccessToken = "string"
)
}
}

\keyword{internal}
