% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_identity_provider_by_identifier}
\alias{cognitoidentityprovider_get_identity_provider_by_identifier}
\title{Gets the specified identity provider}
\usage{
cognitoidentityprovider_get_identity_provider_by_identifier(UserPoolId,
  IdpIdentifier)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{IdpIdentifier}{[required] The identity provider ID.}
}
\description{
Gets the specified identity provider.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$get_identity_provider_by_identifier(
  UserPoolId = "string",
  IdpIdentifier = "string"
)
}
}

\keyword{internal}
