% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_start_user_import_job}
\alias{cognitoidentityprovider_start_user_import_job}
\title{Starts the user import}
\usage{
cognitoidentityprovider_start_user_import_job(UserPoolId, JobId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that the users are being imported
into.}

\item{JobId}{[required] The job ID for the user import job.}
}
\description{
Starts the user import.
}
\section{Request syntax}{
\preformatted{cognitoidentityprovider$start_user_import_job(
  UserPoolId = "string",
  JobId = "string"
)
}
}

\keyword{internal}
